// Cerebral/agent-lee-os2/src/types.ts
export type ShapeType =
  | "ankh"
  | "lotus"
  | "merkaba"
  | "thirdEye"
  | "infinity"
  | "crescent"
  | "pyramid"
  | "sunburst"
  | "sphere"
  | "diamond"
  | "star"
  | "battery"
  | "bear"
  | "heart";

export const SHAPES: ShapeType[] = [
  "ankh",
  "lotus",
  "merkaba",
  "thirdEye",
  "infinity",
  "crescent",
  "pyramid",
  "sunburst",
  "sphere",
  "diamond",
  "star",
  "battery",
  "bear",
  "heart",
];